# public/document/api-doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample web server for Smart-City project.",
    "version": "1.0.0",
    "title": "Smart-City Backend API Doc"
  },
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/point/data/create": {
      "post": {
        "tags": [
          "Point-Data"
        ],
        "operationId": "postPointData",
        "consumes": [
          "application/json",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fileName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "keyPropertyName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "xPropertyName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "yPropertyName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "dataStruct",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "isPublic",
            "in": "formData",
            "required": true,
            "type": "boolean"
          },
          {
            "name": "coordinateSystem",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "formData",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/point/data/list": {
      "get": {
        "tags": [
          "Point-Data"
        ],
        "operationId": "getPointDataList",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
...
    "/api/point/data/read/{fileName}/{version}": {
      "get": {
        "tags": [
          "Point-Data"
        ],
        "operationId": "getPointData",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/point/data/download/{fileName}/{version}": {
      "get": {
        "tags": [
          "Point-Data"
        ],
        "operationId": "downloadPointData",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/point/data/remove/{fileName}/{version}": {
      "delete": {
        "tags": [
          "Point-Data"
        ],
        "operationId": "removePointData",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
...
    "/api/publish": {
      "post": {
        "tags": [
          "Publish"
        ],
        "operationId": "publishService",
        "consumes": [
          "application/json",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "arcgisUser",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "arcgisPassword",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "arcgisPortal",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "serviceName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "fileNames",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "shapefiles",
            "in": "formData",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/publish/apply": {
      "post": {
        "tags": [
          "Publish"
        ],
        "operationId": "applyPublish",
        "consumes": [
          "application/json",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "arcgisUser",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "arcgisPassword",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "arcgisPortal",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "serviceName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "fileNames",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    },
    "/api/publish/status": {
      "get": {
        "tags": [
          "Publish"
        ],
        "operationId": "getPublishStatus",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal exception"
          }
        }
      }
    }
  }
}